refactor: move "recent-fare-contract" as part of "fare-contract" module
