+++ src/modules/fare-contracts/recent/use-recent-fare-contracts-query.ts
